패키지 정리 - 안드로이드 제공 컴포넌트 단위로 패키지화 - Application 클래스 일원화
