+++ app/src/main/java/com/jp_ais_training/keibo/adapter/CalendarItem.kt
@@ -1,0 +1,9 @@
+package com.jp_ais_training.keibo.adapter
+
+data class CalendarItem(
+    val index: Int,
+    val date: String,
+    val income: Int,
+    val expense: Int,
+)
+
