--- app/src/main/java/com/jp_ais_training/keibo/main/home/CalendarItem.kt
@@ -1,9 +1,0 @@
-package com.jp_ais_training.keibo.main.home
-
-data class CalendarItem(
-    val index: Int,
-    val date: String,
-    val income: Int,
-    val expense: Int,
-)
-
