--- app/src/main/java/com/jp_ais_training/keibo/main/model/Entity/MainCategory.kt
@@ -1,12 +1,0 @@
-package com.jp_ais_training.keibo.main.model.Entity
-
-import androidx.room.ColumnInfo
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity(tableName = "MainCategory")
-data class MainCategory(
-    @PrimaryKey val main_category_id: Int,
-    @ColumnInfo(name = "main_category_name") val main_category_name: String,
-    @ColumnInfo(name = "expense_type") val expense_type: String
-)
