+++ app/src/main/res/layout/fragment_bar_statistics.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".main.BarStatisticsFragment">
+    tools:context=".frament.BarStatisticsFragment">
 
     <ScrollView
         android:layout_width="match_parent"
