+++ app/src/main/res/layout/fragment_home.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".main.home.HomeFragment">
+    tools:context=".frament.HomeFragment">
 
     <include
         android:id="@+id/home_calendar"
